requirements.txt
PyQt6==6.6.1
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
openslide-python==1.3.1
scikit-image==0.22.0
scipy==1.11.4
matplotlib==3.8.2
tifffile==2023.9.26